refactor(cli): moved CLI command logic to a separated module
